+++ test/testlock.c
@@ -129,7 +129,7 @@ int main(int argc, char *argv[])
             if (SDL_strcmp(argv[i], "--nbthreads") == 0) {
                 if (argv[i + 1]) {
                     char *endptr;
-                    nb_threads = SDL_strtol(argv[i + 1], &endptr, 0);
+                    nb_threads = (int) SDL_strtol(argv[i + 1], &endptr, 0);
                     if (endptr != argv[i + 1] && *endptr == '\0' && nb_threads > 0) {
                         consumed = 2;
                     }
@@ -137,7 +137,7 @@ int main(int argc, char *argv[])
             } else if (SDL_strcmp(argv[i], "--worktime") == 0) {
                 if (argv[i + 1]) {
                     char *endptr;
-                    nb_threads = SDL_strtol(argv[i + 1], &endptr, 0);
+                    nb_threads = (int) SDL_strtol(argv[i + 1], &endptr, 0);
                     if (endptr != argv[i + 1] && *endptr == '\0' && nb_threads > 0) {
                         consumed = 2;
                     }
@@ -146,7 +146,7 @@ int main(int argc, char *argv[])
             } else if (SDL_strcmp(argv[i], "--timeout") == 0) {
                 if (argv[i + 1]) {
                     char *endptr;
-                    timeout = SDL_strtol(argv[i + 1], &endptr, 0);
+                    timeout = (int) SDL_strtol(argv[i + 1], &endptr, 0);
                     if (endptr != argv[i + 1] && *endptr == '\0' && timeout > 0) {
                         consumed = 2;
                     }
